add docker to project
