+++ package.json
@@ -2,10 +2,10 @@
   "name": "app-backend-api",
   "version": "1.0.0",
   "description": "backend Node API for all my projects",
-  "main": "index.js",
+  "main": "./src/api/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.ts"
+    "start": "nodemon ./src/api/index.ts"
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +18,7 @@
     "@typescript-eslint/eslint-plugin": "^6.11.0",
     "@typescript-eslint/parser": "^6.11.0",
     "eslint": "^8.53.0",
+    "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   }
 }
